Move `MessageType` to its own file.
